backend/requirements.txt
-i https://pypi.org/simple
asgiref==3.8.1; python_version >= '3.8'
certifi==2024.8.30; python_version >= '3.6'
charset-normalizer==3.4.0; python_full_version >= '3.7.0'
django==5.1.3; python_version >= '3.10'
django-cors-headers==4.6.0; python_version >= '3.9'
django-material-admin==1.8.6
django-rest-framework==0.1.0
djangorestframework==3.15.2; python_version >= '3.8'
djangorestframework-simplejwt==5.3.1; python_version >= '3.8'
djongo==1.2.31; python_version >= '3.6'
idna==3.10; python_version >= '3.6'
psycopg2==2.9.10; python_version >= '3.8'
psycopg2-binary==2.9.10; python_version >= '3.8'
pyjwt==2.9.0; python_version >= '3.8'
pymongo==4.10.1; python_version >= '3.8'
python-dotenv==1.0.1; python_version >= '3.8'
pytz==2024.2
requests==2.32.3; python_version >= '3.8'
sqlparse==0.5.2; python_version >= '3.8'
urllib3==2.2.3; python_version >= '3.8'
